requirements.txt
alembic==1.13.2
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
asttokens==2.4.1
attrs==24.2.0
bcrypt==4.2.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.4
fastapi-cli==0.0.5
fastjsonschema==2.21.1
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.10.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.1.2
openpyxl==3.1.5
orjson==3.10.12
packaging==24.1
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.1
pydantic-extra-types==2.10.0
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
seaborn==0.13.2
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.40.0
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.1
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
websockets==14.1
wheel==0.44.0
zipcodes-in==0.2.0
